update wwith database working

+++ Breathe/app/src/main/java/com/example/breathe/data/models/MeditationSession.kt
@@ -5,6 +5,6 @@ import androidx.room.PrimaryKey
 @Entity(tableName = "meditation_sessions")
 data class MeditationSession(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
-    val duration: Long, // длительность в секундах
+    val duration: Int, // длительность в секундах
     val date: Long     // дата в формате timestamp
 )
